Fix android corpses
Inorganic androids can have the nocorpse flag.

Parasitized androids have a new flag, spcorpse, that blocks corpses of that type but doesn't have the other nocorpse behavior (like being worth 0 nutrition).
-Up the size of he monster gen flags to allow this new flag.

+++ include/decl.h
@@ -548,7 +548,7 @@ E NEARDATA struct permonst upermonst;	/* init'd in decl.c,
 E NEARDATA struct mvitals {
 	uchar	born;
 	uchar	died;
-	uchar	mvflags;
+	long long mvflags;
 	int	san_lost;
 	int	insight_gained;
 	Bitfield(seen,1);
